finished running go fmt command on each file, reviewing formatting, editing comments, updating readme, logs, etc.
