+++ cmd/cmd-runner.go
@@ -1,9 +1,10 @@
 package cmd
 
 import (
-	"CacheCLI/kvcache"
 	"fmt"
 	"github.com/spf13/cobra"
+	
+	"CacheCLI/kvcache"
 )
 
 type CommandRunner struct {
