+++ kvcache/mock_kvcache.go
@@ -5,22 +5,13 @@ import (
 	"strconv"
 )
 
-/* MockCache struct that is a member of the  KVC interface for testing of KVC CLI commands */
+/* implementation of the KeyValueCache interface for testing */
 type MockKeyValueCache struct {
 	Success      bool
 	ReturnString string
 }
 
-//go insists that the put method is required to implement the KVC interface for the NewMockSimpleKVCache constructor.
-// ..I've invalidated the files/cache several times and it still requires this - so putting it in for now but NOT
-// USED - can't find it's use ANYWHERE
-func (m *MockKeyValueCache) Put(key, value string) error {
-	m.Success, _ = strconv.ParseBool(key)
-	m.ReturnString = value
-	return nil
-}
-
-//constructor function for generating test MockCache
+//constructor function useful for testing
 func NewMockSimpleKVCache(success bool, retString string) KeyValueCache {
 	return &MockKeyValueCache{success, retString}
 }
